1.0.6 	modified:   Makefile 	new file:   New Text Document.txt 	modified:   src/execute/execute.c 	modified:   src/execute/execute.h 	modified:   src/exeviewer/exeviewer.c 	modified:   src/exeviewer/exeviewer.h 	modified:   src/file/fileread.c 	modified:   src/formats/win-exe/Makefile 	new file:   src/formats/win-exe/winexe.c 	modified:   src/formats/win-exe/winexe_functions.c 	modified:   src/formats/win-exe/winexe_functions.h 	modified:   src/formats/win-exe/winexe_struct_filler_functions.c 	modified:   src/formats/win-exe/winexe_struct_filler_functions.h 	modified:   src/formats/win-exe/winexe_structures.h 	modified:   src/main.c 	modified:   src/types.h

+++ src/types.h
@@ -35,8 +35,8 @@ typedef unsigned __int64  ULONG_PTR;
 // Windows-compatible macros.
 #define MAKEWORD(a,b) ((WORD) (((BYTE)  (((DWORD_PTR) (a)) & 0xff)) | ((WORD) ((BYTE) (((DWORD_PTR) (b)) & 0xff))) << 8))
 #define MAKELONG(a,b) ((LONG) (((WORD)  (((DWORD_PTR) (a)) & 0xffff)) | ((DWORD) ((WORD) (((DWORD_PTR) (b)) & 0xffff))) << 16))
-#define MAKELONGLONG(a,b)  ((LONGLONG)  ((((DWORD) (a)) & 0xffffffff) | ((((DWORD) (b)) & 0xffffffff)) << 32))
-#define MAKEULONGLONG(a,b) ((ULONGLONG) ((((DWORD) (a)) & 0xffffffff) | ((((DWORD) (b)) & 0xffffffff)) << 32))
+#define MAKELONGLONG(a,b)  ((LONGLONG)  ((((DWORD) (a)) & 0xffffffff) | (LONGLONG)((((DWORD) (b)) & 0xffffffff)) << 32))
+#define MAKEULONGLONG(a,b) ((ULONGLONG) ((((DWORD) (a)) & 0xffffffff) | (ULONGLONG)((((DWORD) (b)) & 0xffffffff)) << 32))
